refactor wasm memory demands validation to a separate file :shrug: :facepalm:
