+++ cosmwasm/packages/wasmi-runtime/src/wasm/contract_operations.rs
@@ -15,7 +15,10 @@ use super::contract_validation::{
 use super::errors::wasmi_error_to_enclave_error;
 use super::gas::{gas_rules, WasmCosts};
 use super::io::encrypt_output;
-use super::runtime::{create_builder, ContractInstance, Engine, WasmiImportResolver};
+use super::{
+    memory::validate_memory,
+    runtime::{create_builder, ContractInstance, Engine, WasmiImportResolver},
+};
 use crate::wasm::types::SecretMessage;
 
 /*
@@ -220,51 +223,9 @@ fn start_engine(
 
     trace!("Deserialized Wasm contract");
 
-    let memory_section = p_modlue
-        .memory_section_mut()
-        .ok_or(EnclaveError::CannotInitializeWasmMemory).map_err(|err| {
-            error!(
-                "Error accessing memory section of WASM while trying to validate memory demands: {:?}",
-                err
-            );
-            err
-        })?;
+    trace!("Validating WASM memory demands");
 
-    if memory_section.entries().len() != 1 {
-        error!(
-            "WASM demands too many memory sections. Must be 1, demands {}",
-            memory_section.entries().len()
-        );
-        return Err(EnclaveError::CannotInitializeWasmMemory);
-    }
-
-    let memory_entry = memory_section
-        .entries()
-        .first()
-        .ok_or(EnclaveError::CannotInitializeWasmMemory)
-        .map_err(|err| {
-            error!(
-                "Error accessing memory entry of WASM while trying to validate memory demands: {:?}",
-                err
-            );
-            err
-        })?;
-
-    let requested_initial_pages: u32 = memory_entry.limits().initial();
-    let maximum_allowed_pages: u32 = 192; // 12 MiB
-
-    if requested_initial_pages > maximum_allowed_pages {
-        error!(
-            "WASM Requested to initialize with {} pages, maximum allowed is {}",
-            requested_initial_pages, maximum_allowed_pages
-        );
-        return Err(EnclaveError::CannotInitializeWasmMemory);
-    }
-
-    *memory_section.entries_mut() = vec![MemoryType::new(
-        requested_initial_pages,
-        Some(maximum_allowed_pages),
-    )];
+    validate_memory(&mut p_modlue)?;
 
     trace!("Validated WASM memory demands");
