+++ cosmwasm/packages/wasmi-runtime/src/wasm/mod.rs
@@ -1,11 +1,11 @@
 mod contract_operations;
 mod contract_validation;
+mod db;
 mod errors;
 mod gas;
-mod runtime;
-
-mod db;
 mod io;
+mod memory;
+mod runtime;
 mod types;
 
 pub use contract_operations::{handle, init, query};
